student_vaccination_app/frontend/requirements.txt
cryptography 
Flask==2.0.2
Flask-MySQL==1.5.1
mysql-connector-python
